Move makeNamespacedInputCopy to separate ts file
