+++ packages/core/src/utils/make-namespaced.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `src/storage/namespace.test.ts`
+# Snapshot report for `src/utils/make-namespaced.test.ts`
 
-The actual snapshot is saved in `namespace.test.ts.snap`.
+The actual snapshot is saved in `make-namespaced.test.ts.snap`.
 
 Generated by [AVA](https://avajs.dev).
